// rss-css-selectors/src/components/view/sidebar/enum.ts
enum STYLES {
  Sidebar = 'sidebar',
  SidebarOpen = 'sidebar--open',
  TaskTitle = 'task__title',
  TaskDesc = 'task__description',
  TaskExample = 'task__example',
  ExampleTitle = 'task__example-title',
  PanelContainer = 'panel-container',
  Panel = 'panel',
  PanelTitle = 'panel__title',
  LvlStatus = 'level-status',
  HintStatus = 'hint-status',
  LvlStatusActive = 'level-status--active',
  HintStatusActive = 'hint-status--active',
  Control = 'control',
  BtnWrapper = 'btn-wrapper',
  BurgerActive = 'btn-wrapper--active',
  ControlBtnPrev = 'control__btn-prev',
  ControlBtnNext = 'control__btn-next',
  ControlBtnMenu = 'control__btn-menu',
  ControlBtnMain = 'btn-open',
  ControlBtnMainActive = 'btn-open--active',
  BtnReset = 'btn-reset',
  ProgressBar = 'progress-bar',
  Levels = 'levels',
  LevelsTitle = 'levels__title',
  LevelsList = 'levels__list',
  LevelsItem = 'levels__item',
  LevelsItemName = 'levels__item-name',
  LevelsItemSelected = 'levels__item-selected',
  LevelsOpen = 'levels--open',
}

enum TAGNAMES {
  Aside = 'aside',
  H3 = 'h3',
  H4 = 'h4',
  H5 = 'h5',
  P = 'p',
  TaskExample = 'p',
  Div = 'div',
  Progress = 'progress',
  Ol = 'ol',
  Li = 'li',
}

enum CONTENT {
  BtnReset = 'Reset Progress',
  LevelsTitle = 'Choose a level',
}

enum EVENTS {
  Click = 'click',
}

export { STYLES, TAGNAMES, CONTENT, EVENTS };